Include corners in adjacency calculations
The app used to include corners when counting the number of adjacent tiles, but
I removed this so that only tiles directly to the left, top, right, or bottom
counted. I have reversed course because the corners do provide extra weight
to a tile's "stickiness" on the board - it should be quite rare to move a tile
that is completely surrounded by other tiles of its own department.