Encapsulate adjacent / empty tile search order inside Termites class
The act of looking for where to place a tile used to be covered by the
LookAround method in Form1. Its original intent to establish randomly
ordered pairs representing all the tiles immediately surrounding and including
a given position. It did not accomplish this - I discovered that the method
allowed duplicate positions to creep in. This actually made the program work,
though. Not checking all tiles around the termite's position made the tiles
more mobile, allowing small clusters of identical tiles to eventually be broken
up and brought into the main cluster.

I have, therefore, decided to replicate the bug by having each termite check a
randomly determined number of tiles around it when moving around. I'be moved
away from using a 2D array for this, preferring instead to use a list of
Positions. I refactored the GreedyTermite and ScholarTermite methods in Form1
to make use of the new objects.