Subclass Termites to Greedy and Scholar termites
The distinction between a greedy and scholar termite was originally made with
the flag 'TermiteType' on the Termites class. I removed this and made two
proper subclasses of Termites: ScholarTermite and GreedyTermite. I moved their
namesake methods from form1 into this class, and added an event to the
FacilityLayoutModel so that the UI could subscribe and update tiles
appropriately. I was able to remove a great deal of redundant code from Form1.