Port everything related to myTermiteOwnedTiles to C#
Added a LockedTile array to FacilityLayoutModel, which better expresses the
usage of myTermiteOwnedTile. Added utility methods to handle locking/unlocking
groups of tiles.

Add TakeTile method to Termites. This handles the act of setting the termite's
HasTile property to true, setting the termite's tile dept, and removing the
tile from the FacilityLayoutModel, which in turn raises an event for the UI.